chat refactor1
