+++ README.md
@@ -12,10 +12,11 @@ A monorepo for personal learning projects and production AI-based services for m
     * Mobile: 
         * Framework: Expo
         * UI: TBD - [Expo UI (Beta)](https://docs.expo.dev/versions/v53.0.0/sdk/ui/)? [React Native Reusables](https://rnr-docs.vercel.app/getting-started/introduction/)? Want something that can render on ios, android, and web.
+    * Package Manager: pnpm
 * Backend: [FastAPI](https://fastapi.tiangolo.com/)
     * Database: [Neon Postgres](https://neon.tech/)
+    * Package Manager: uv
 * Deployment: [Railway](https://railway.com/project/73eb837a-ba86-4899-992c-cefd0c22b91f?environmentId=455c3498-682b-4e4d-9e1f-4c13c3e9eb59)
-* Package Managers: pnpm and uv
 
 ## Setup
