+++ pyproject.toml
@@ -63,4 +63,5 @@ dependencies = [
 [dependency-groups]
 dev = [
     "ipython>=9.2.0",
+    "ipykernel",
 ]
